+++ vanilla-redux/public/index.html
@@ -14,8 +14,11 @@
     <title>Vanilla Redux</title>
   </head>
   <body>
-    <button id="add">Add</button>
-    <span></span>
-    <button id="minus">Minus</button>
+    <h1>To Dos</h1>
+    <form>
+      <input type="text" placeholder="Write to do" />
+      <button>Add</button>
+    </form>
+    <ul></ul>
   </body>
 </html>
