+++ vanilla-redux/src/index.js
@@ -1,35 +1,32 @@
-import { legacy_createStore } from "redux";
+import {legacy_createStore} from "redux"; 
 
-const add = document.getElementById("add");
-const minus = document.getElementById("minus");
-const number = document.querySelector("span");
+const form = document.querySelector("form");
+const input = document.querySelector("input");
+const ul = document.querySelector("ul");
 
-number.innerText = 0;
+const ADD_TODO = "ADD_TODO";
+const DELETE_TODO = "DELETE_TODO";
 
-const ADD = "ADD";
-const MINUS = "MINUS"
-
-const countModifier = (count = 0, action) => {
-    switch (action.type){
-        case ADD:
-            return count+1;
-        case MINUS:
-            return count-1;
+const reducer = (state=[], action)=>{
+    console.log(action);
+    switch(action.type){
+        case ADD_TODO:
+            return [...state, {text: action.text, id: Date.now()}];
+        case DELETE_TODO:
+            return[];
         default:
-            return count;
+           return state;
     }
 };
 
-const countStore = legacy_createStore(countModifier);
+const store = legacy_createStore(reducer);
 
-const onChange = ()=>{
-    number.innerText = countStore.getState();
-}
+store.subscribe(()=> console.log(store.getState()));
+const onSubmit = e => {
+e.preventDefault();
+const toDo = input.value;
+input.value = "";
+store.dispatch({type:ADD_TODO, text: toDo});
+};
 
-countStore.subscribe(onChange);
-
-const handleAdd =()=>{countStore.dispatch({type:ADD})}
-const handleMinus =()=>{countStore.dispatch({type:MINUS})}
-
-add.addEventListener("click", handleAdd);
-minus.addEventListener("click", handleMinus);
+form.addEventListener("submit", onSubmit);
